dataset/requirements.txt
datasets>=2.14.0
huggingface_hub>=0.17.0
pyarrow>=12.0.0
pandas>=2.0.0
python-dotenv>=1.0.0
